refactor: :truck: Results Move
Se quito el result de application y se coloco en una carpeta common porque no es parte de ninguna de las carpetas anteriores y por ende se modifico en todos los archivos en el commit
