+++ lib/domain/repositories/client/client_repository.dart
@@ -1,4 +1,4 @@
-import 'package:gymnastic_center/application/core/results.dart';
+import 'package:gymnastic_center/common/results.dart';
 import 'package:gymnastic_center/domain/entities/client/client.dart';
 
 abstract class ClientRepository {
